Add elapsed time print
