+++ tst/ogb/main_pyg_with_pruning.py
@@ -2,11 +2,9 @@ from time import time
 
 import argparse
 from functools import partial
-from pathlib import Path
 from typing import Dict
 
 import numpy as np
-import pandas as pd
 import torch
 import torch.optim as optim
 from torch_geometric.data import DataLoader
@@ -14,13 +12,12 @@ from torchvision import transforms
 from tqdm import tqdm
 
 from ogb.graphproppred import PygGraphPropPredDataset, Evaluator
-from ogb.graphproppred.mol_encoder import AtomEncoder, BondEncoder
 from src.utils.graph_prune_utils import tg_dataset_prune
 from src.utils.lsh_euclidean_tools import LSH
 from src.utils.minhash_tools import MinHash
 from src.utils.proxy_utils import set_proxy
-from tst.ogb.encoder_utils import ASTNodeEncoder, augment_edge, decode_arr_to_seq, encode_y_to_arr, get_vocab_mapping
-from tst.ogb.gcn import GCN
+from tst.ogb.encoder_utils import augment_edge, decode_arr_to_seq, encode_y_to_arr, get_vocab_mapping
+from tst.ogb.model_and_data_utils import add_zeros, create_model
 
 
 def train(model, device, loader, optimizer, cls_criterion):
@@ -90,35 +87,6 @@ def evaluate(model, device, loader, evaluator, arr_to_seq, dataset_name: str):
         input_dict = {"y_true": y_true, "y_pred": y_pred}
 
     return evaluator.eval(input_dict)
-
-
-def create_model(dataset: PygGraphPropPredDataset, emb_dim: int, dropout_ratio: float, device: str, num_layers: int,
-                 max_seq_len: int, num_vocab: int):
-    print("creating a model for ", dataset.name)
-    if dataset.name == "ogbg-molhiv":
-        node_encoder = AtomEncoder(emb_dim=emb_dim)
-        edge_encoder_constrtuctor = BondEncoder
-        print("Number of classes: ", dataset.num_tasks)
-        model = GCN(num_classes=dataset.num_tasks, num_layer=num_layers,
-                    emb_dim=emb_dim, drop_ratio=dropout_ratio,
-                    node_encoder=node_encoder, edge_encoder_ctor=edge_encoder_constrtuctor).to(device)
-
-    elif dataset.name == "ogbg-code":
-        nodetypes_mapping = pd.read_csv(Path(dataset.root) / 'mapping' / 'typeidx2type.csv.gz')
-        nodeattributes_mapping = pd.read_csv(Path(dataset.root) / 'mapping' / 'attridx2attr.csv.gz')
-        node_encoder = ASTNodeEncoder(emb_dim, num_nodetypes=len(nodetypes_mapping['type']),
-                                      num_nodeattributes=len(nodeattributes_mapping['attr']), max_depth=20)
-        split_idx = dataset.get_idx_split()
-        vocab2idx, idx2vocab = get_vocab_mapping([dataset.data.y[i] for i in split_idx['train']], num_vocab)
-        edge_encoder_ctor = partial(torch.nn.Linear, 2)
-        print(f"Multiclassification with {len(vocab2idx)} classes. Num labels per example: {max_seq_len}")
-        model = GCN(num_classes=len(vocab2idx), max_seq_len=max_seq_len, node_encoder=node_encoder,
-                    edge_encoder_ctor=edge_encoder_ctor, num_layer=num_layers, emb_dim=emb_dim,
-                    drop_ratio=dropout_ratio).to(device)
-
-    else:
-        raise ValueError("Used an invalid dataset name")
-    return model
 
 
 def get_prune_args(pruning_method: str, num_minhash_funcs: int, random_pruning_prob: float, node_dim: int) -> Dict:
@@ -193,7 +161,8 @@ def main():
         set_proxy()
 
     # automatic data loading and splitting
-    dataset = PygGraphPropPredDataset(name=args.dataset)
+    transform = add_zeros if args.dataset == 'ogbg-ppa' else None
+    dataset = PygGraphPropPredDataset(name=args.dataset, transform=transform)
     cls_criterion = torch.nn.CrossEntropyLoss() if args.dataset == 'ogbg-code' else torch.nn.BCEWithLogitsLoss()
     idx2word_mapper = None
     split_idx = dataset.get_idx_split()
